Slots for 'upstream' and 'downstream' changed from numeric to integer. writeVcf() bugs now fixed, all examples are running.
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/VariantAnnotation@68819 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/AllClasses.R
@@ -140,8 +140,8 @@ setClass("SpliceSiteVariants", contains="VariantType")
 
 setClass("PromoterVariants", 
     contains="VariantType",
-    representation(upstream="numeric",
-                   downstream="numeric")
+    representation(upstream="integer",
+                   downstream="integer")
 )
 
 .promoterValidity <- function(object, ...)
@@ -158,8 +158,8 @@ setValidity("PromoterVariants",
 
 setClass("AllVariants", 
     contains="VariantType",
-    representation(upstream="numeric",
-                   downstream="numeric")
+    representation(upstream="integer",
+                   downstream="integer")
 )
 
 setValidity("AllVariants",
